requirements.txt
google-api-python-client
pyzmq
google-cloud-firestore